Fix padding on some art files (See Commit Description)
I've been working on a byte perfect compressor for the proprietary art compression used in this game, and it's nearly ready to be implemented into the repo.  The plan is to store all of the art assets as uncompressed in the repo so that they can be edited more easily and compress them on build time, but I needed to fix the padding on some of the files first.